conntrackd: add libsystemd integration
A couple of patches are added to give conntrackd libsystemd integration.

Now, the admin can choose Type=notify (our default starting now) to get a
better interaction between systemd and conntrackd.
Both patches are already in their way upstream.

This also enables systemd Watchdog support, and the conntrackd.service file
includes the directive commented as a hint for the admin.

Signed-off-by: Arturo Borrero Gonzalez <arturo.borrero.glez@gmail.com>